backend/requirements.txt
fastapi
firebase-admin
googletrans
requests
uvicorn
google
google-genai
google-generativeai
python-dateutil
twilio
python-dotenv
urllib3
python-multipart